undo temp and get_child changes
