+++ extra/models/mask_rcnn.py
@@ -5,7 +5,8 @@ import numpy as np
 from pathlib import Path
 from tinygrad import nn
 from tinygrad.tensor import Tensor
-from tinygrad.helpers import dtypes, fetch, get_child
+from extra.utils import get_child
+from tinygrad.helpers import dtypes, fetch
 from tinygrad.nn.state import torch_load
 from extra.models.resnet import ResNet
 from extra.models.retinanet import nms as _box_nms
