+++ extra/models/retinanet.py
@@ -1,7 +1,8 @@
 import math
-from tinygrad.helpers import flatten, get_child
+from tinygrad.helpers import flatten
 import tinygrad.nn as nn
 from extra.models.resnet import ResNet
+from extra.utils import get_child
 import numpy as np
 
 def nms(boxes, scores, thresh=0.5):
