+++ test/models/test_onnx.py
@@ -5,9 +5,10 @@ import io
 import unittest
 import numpy as np
 import onnx
+from extra.utils import temp
 from extra.onnx import get_run_onnx
 from tinygrad.tensor import Tensor
-from tinygrad.helpers import CI, fetch, temp
+from tinygrad.helpers import CI, fetch
 import pytest
 
 pytestmark = [pytest.mark.exclude_gpu, pytest.mark.exclude_clang]
